requirements.txt
langchain
langchain-google-genai
langgraph
chromadb
sentence-transformers
streamlit
pypdf
datasets
python-dotenv
